+++ src/infrastructure/http/interfaces/RequestOptions.ts
@@ -1,5 +1,6 @@
 export interface RequestOptions {
   method: string
   headers: Record<string, string>
+  credentials?:  "omit" | "same-origin" | "include"
   body?: string
 }