+++ src/services/userActivity.js
@@ -13,7 +13,7 @@ import {
 import { DataContext, UserActivityVersionKey } from './dataContext.js'
 import { fetchByRailContentId, fetchByRailContentIds, fetchShows } from './sanity'
 import { fetchPlaylist, fetchUserPlaylists } from './content-org/playlists'
-import { pinnedGuidedCourses } from './content-org/guided-courses'
+import {guidedCourses, pinnedGuidedCourses} from './content-org/guided-courses'
 import {
   getMonday,
   getWeekNumber,
@@ -995,6 +995,7 @@ function generateContentsMap(contents, playlistsContents) {
   const existingShows = new Set()
   const contentsMap = new Map()
   const childToParentMap = {}
+  if (!contents) return contentsMap
   contents.forEach((content) => {
     if (Array.isArray(content.parent_content_data) && content.parent_content_data.length > 0) {
       childToParentMap[content.id] =
@@ -1050,13 +1051,19 @@ function generateContentsMap(contents, playlistsContents) {
  */
 export async function getProgressRows({ brand = null, limit = 8 } = {}) {
   // TODO slice progress to a reasonable number, say 100
-  const [recentPlaylists, progressContents, allPinnedGuidedCourse, userPinnedItem] =
+
+  const [recentPlaylists, progressContents, allPinnedGuidedCourse, userPinnedItem, enrolledGuidedCourses] =
     await Promise.all([
       fetchUserPlaylists(brand, { sort: '-last_progress', limit: limit }),
       getAllStartedOrCompleted({ onlyIds: false, brand: brand }),
       pinnedGuidedCourses(brand),
       getUserPinnedItem(brand),
+      guidedCourses()
     ])
+
+  const enrolledGuidedCoursesIds = enrolledGuidedCourses.map(course => String(course.content_id));
+
+  const mergedGuidedCourses = 1 //get all cuigded courses, incl no progress.
   let pinnedGuidedCourse = allPinnedGuidedCourse?.[0] ?? null
 
   const playlists = recentPlaylists?.data || []
@@ -1066,27 +1073,30 @@ export async function getProgressRows({ brand = null, limit = 8 } = {}) {
   )
 
   const nonPlaylistContentIds = Object.keys(progressContents)
-  if (pinnedGuidedCourse) {
-    nonPlaylistContentIds.push(pinnedGuidedCourse.content_id)
+  if (enrolledGuidedCoursesIds.length > 0) {
+    nonPlaylistContentIds.push(...enrolledGuidedCoursesIds)
   }
   if (userPinnedItem?.progressType === 'content') {
     nonPlaylistContentIds.push(userPinnedItem.id)
   }
+console.log("nonPlaylistContentIds",nonPlaylistContentIds)
+
+
   const [playlistsContents, contents] = await Promise.all([
-    addContextToContent(fetchByRailContentIds, playlistEngagedOnContents, 'progress-tracker', {
+    playlistEngagedOnContents ? addContextToContent(fetchByRailContentIds, playlistEngagedOnContents, 'progress-tracker', {
       addNextLesson: true,
       addNavigateTo: true,
       addProgressStatus: true,
       addProgressPercentage: true,
       addProgressTimestamp: true,
-    }),
-    addContextToContent(fetchByRailContentIds, nonPlaylistContentIds, 'progress-tracker', brand, {
+    }) : Promise.resolve([]),
+    nonPlaylistContentIds ? addContextToContent(fetchByRailContentIds, nonPlaylistContentIds, 'progress-tracker', brand, {
       addNextLesson: true,
       addNavigateTo: true,
       addProgressStatus: true,
       addProgressPercentage: true,
       addProgressTimestamp: true,
-    }),
+    }) : Promise.resolve([]),
   ])
   const contentsMap = generateContentsMap(contents, playlistsContents)
   let combined = await extractPinnedItemsAndSortAllItems(
@@ -1103,7 +1113,6 @@ export async function getProgressRows({ brand = null, limit = 8 } = {}) {
         item.type === 'playlist' ? processPlaylistItem(item) : processContentItem(item)
       )
   )
-  console.log('HomePageProgressRows results: remove before merge', results)
   return {
     type: TabResponseType.PROGRESS_ROWS,
     displayBrowseAll: combined.length > limit,